CIA-24 Move tippy defs into typings.d.ts
